Add list of all available permissions
