+++ src/components/permission/PermissionSpec.tsx
@@ -1,0 +1,140 @@
+import { CommandData, PermissionData } from "@site/src/components/permission/Permissions";
+
+export const vanillaCommands: CommandData = {
+  "0": "minecraft.command",
+  "1": [
+    ["advancement", [], 2],
+    ["attribute", [], 2],
+    ["ban", [], 3],
+    ["ban-ip", [], 3],
+    ["banlist", [], 3],
+    ["bossbar", [], 2],
+    ["clear", [], 2],
+    ["clone", [], 2],
+    ["damage", [], 2],
+    ["data", [], 2],
+    ["datapack", [], 2],
+    ["debug", [], 3],
+    ["defaultgamemode", [], 2],
+    ["deop", [], 3],
+    ["difficulty", [], 2],
+    ["effect", [], 2],
+    ["enchant", [], 2],
+    ["execute", [], 2],
+    ["fill", [], 2],
+    ["fillbiome", [], 2],
+    ["forceload", [], 2],
+    ["function", [], 2],
+    ["gamemode", [], 2],
+    ["gamerule", [], 2],
+    ["give", [], 2],
+    ["help", [], 0],
+    ["item", [], 2],
+    ["jfr", [], 4],
+    ["kick", [], 3],
+    ["kill", [], 2],
+    ["list", [], 0],
+    ["locate", [], 2],
+    ["loot", [], 2],
+    ["me", [], 0],
+    ["msg", ["w", "tell"], 0],
+    ["op", [], 3],
+    ["pardon", [], 3],
+    ["pardon-ip", [], 3],
+    ["particle", [], 2],
+    ["perf", [], 4],
+    ["place", [], 2],
+    ["playsound", [], 2],
+    ["publish", [], 4],
+    ["random", [], "Yes without sequence argument, No with sequence argument."],
+    ["recipe", [], 2],
+    ["reload", [], 2],
+    ["return", [], 2],
+    ["ride", [], 2],
+    ["save-all", [], 4],
+    ["save-off", [], 4],
+    ["save-on", [], 4],
+    ["say", [], 2],
+    ["schedule", [], 2],
+    ["scoreboard", [], 2],
+    ["seed", [], 2],
+    ["setblock", [], 2],
+    ["setidletimeout", [], 3],
+    ["setworldspawn", [], 2],
+    ["spawnpoint", [], 2],
+    ["spectate", [], 2],
+    ["spreadplayers", [], 2],
+    ["stop", [], 4],
+    ["stopsound", [], 2],
+    ["summon", [], 2],
+    ["tag", [], 2],
+    ["team", [], 2],
+    ["teammsg", ["tm"], 0],
+    ["teleport", ["tp"], 2],
+    ["tellraw", [], 2],
+    ["tick", [], 3],
+    ["time", [], 2],
+    ["title", [], 2],
+    ["transfer", [], 3],
+    ["trigger", [], 0],
+    ["weather", [], 2],
+    ["whitelist", [], 3],
+    ["worldborder", [], 2],
+    ["xp", ["experience"], 2],
+  ],
+};
+
+export const bukkitCommands: CommandData = {
+  "0": "bukkit.command",
+  "1": [
+    ["help", ["?"], 0],
+    ["plugins", ["pl"], 0],
+    ["reload", ["rl"], 4],
+    ["timings", [], 4],
+    ["version", ["ver", "about"], 0],
+  ],
+};
+
+export const paperCommands: CommandData = {
+  "0": "bukkit.command",
+  "1": [["paper", [], 4]],
+};
+
+export const vanillaPermissions: PermissionData = {
+  "0": "minecraft",
+  "1": [
+    ["", 4],
+    ["admin.command_feedback", 4],
+    ["nbt.place", 4],
+    ["nbt.copy", 0],
+    ["debugstick", 4],
+    ["debugstick.always", false],
+    ["commandblock", 4],
+  ],
+};
+
+export const bukkitPermissions: PermissionData = {
+  "0": "bukkit",
+  "1": [
+    ["broadcast", 4],
+    ["broadcast.admin", 4],
+    ["broadcast.user", 0],
+  ],
+};
+
+export const paperPermissions: PermissionData = {
+  "0": "bukkit.command.paper",
+  "1": [
+    ["heap", 4],
+    ["entity", 4],
+    ["reload", 4],
+    ["version", 4],
+    ["dumpplugins", 4],
+    ["syncloadinfo", 4],
+    ["dumpitem", 4],
+    ["mobcaps", 4],
+    ["dumplisteners", 4],
+    ["fixlight", 4],
+    ["debug", 4],
+  ],
+};
